Add files via upload
This is an example of when I need to pull back and redo my code.  My logic on the original was bad because I messed up my validation check and put the loop too deep.  Because of this, the process never successfully looped for repeated entries.  I decided to step back, pull everything up a level, and the process worked correctly this time.